feat(production): socle #170 - patch OF idempotence/immuabilite + moteur generation unique deplace en domaine production
